requirements.txt
# Telegram Bot
aiogram>=3.10.0

# Database (Postgres)
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis & Message Queue
redis>=5.0.0
aio-pika>=9.4.0  # Для RabbitMQ

# Scheduler & Google
APScheduler>=3.10.0
gspread-asyncio>=2.0.0
google-auth>=2.29.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
structlog>=24.1.0  # JSON логирование
python-dotenv>=1.0.0
aiohttp>=3.9.0

openai>=1.0.0
httpx>=0.27.0